// app/frontend/src/context/api/models/TodoState.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HiveSync API 
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const TodoState = {
    ToDo: 'ToDo',
    InProgress: 'InProgress',
    Done: 'Done'
} as const;
export type TodoState = typeof TodoState[keyof typeof TodoState];


export function instanceOfTodoState(value: any): boolean {
    for (const key in TodoState) {
        if (Object.prototype.hasOwnProperty.call(TodoState, key)) {
            if (TodoState[key as keyof typeof TodoState] === value) {
                return true;
            }
        }
    }
    return false;
}

export function TodoStateFromJSON(json: any): TodoState {
    return TodoStateFromJSONTyped(json, false);
}

export function TodoStateFromJSONTyped(json: any, ignoreDiscriminator: boolean): TodoState {
    return json as TodoState;
}

export function TodoStateToJSON(value?: TodoState | null): any {
    return value as any;
}

export function TodoStateToJSONTyped(value: any, ignoreDiscriminator: boolean): TodoState {
    return value as TodoState;
}